Training data set
The dataset includes detailed information about each vehicle, such as its make, model, year, odometer reading, auction price, and warranty costs, among other attributes. This data provides a solid foundation for the initial exploratory data analysis (EDA) and training the model to predict "kick" cars.

The IsBadBuy column is the target variable, indicating whether a vehicle was a bad buy (kick), with 1 representing a bad buy and 0 representing a good buy. The goal is to build a model that can accurately predict the likelihood of a vehicle being a kick based on these features.

During EDA, the goal is to understand the relationships between features and identify important patterns or trends that help differentiate a good buy from a bad buy.